chore: update to 1.21.10 (+change text display opacity to Int)

+++ src/main/kotlin/dev/znci/rocket/scripting/globals/tables/entities/TextDisplay.kt
@@ -33,8 +33,8 @@ class TextDisplay(entity: TextDisplay): BaseDisplayEntity<TextDisplay>(entity) {
         set(value) { entity.backgroundColor = value.toJava() }
 
     @TwineNativeProperty
-    var textOpacity: Double
-        get() = entity.textOpacity.toDouble()
+    var textOpacity: Int
+        get() = entity.textOpacity.toInt()
         set(value) { entity.textOpacity = value.toInt().toByte() }
 
     @TwineNativeProperty
